gfs2_tool, libgfs2: bz487608 - GFS2: gfs2_tool unfreeze hangs
gfs2_tool tries to stat() the mountpoint in order to get to the
device number and ultimately get to the sysfs freeze tunable.

When the filesystem is frozen, followed by another process holding
an exclusive lock on the mountpoint (eg. touch creating a new file
in root dir), gfs2_tool hangs behind this lock at the stat() call.

This patch makes freeze/unfreeze stat the block device instead of
the mountpoint.

Signed-off-by: Abhijith Das <adas@redhat.com>