gfs2_tool: gettext support
This is only the basics. Message catalog generation will come
later.

My plan is to add gettext support to all the programs, but not
to libgfs2. My intended solution for libgfs2 is to simply avoid
the generation of printable strings and move all interaction with
the user into the programs.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>