libgfs2: Remove die from compute_heightsize
This patch removes the call to die from compute_heightsize and adds
error reporting. To do this the prototype had to be changed to return
the max height value through a pointer so that the return value could
indicate success or failure. The compute_constants prototype was updated
similarly as it calls compute_heightsize. All calls to both functions
are updated to check for errors and print the same error message as before.

Signed-off-by: Andrew Price <andy@andrewprice.me.uk>