gfs2_edit savemeta needs to save freemeta blocks
bz 502056

The gfs2_edit savemeta function was not saving off
blocks of type 2, which correspond to "free metadata" in
gfs and "unlinked dinodes" in gfs2.  This turned out to be
a problem because if that metadata is restored and gfs is
tested against it, it resulted in metadata allocation
problems.  For example, gfs_grow may make the decision to
allocate a metadata block, and try to re-use a block that's
marked as "free metadata" in the gfs bitmap.  But since the
block hadn't been saved at "savemeta" time, the block was
never restored as metadata and therefore the metadata
header may be missing, so gfs-kernel withdrew with a file
system error.