gfs2_tool: Remove code to read args/*
Nobody has replied to my message about removing the printing
of the args parameters in the df subcommand of gfs2_tool. In
any case, all the information is available via /proc/mounts and
that will actually be both correct and complete unlike that in
the args subdirectory of gfs2's sysfs.

In addition there is really no need to list this information
during df since its not likely to be of any interest to someone
who is only interested in knowing how full their filesystem is.

Also there was an error which has been around for some time
in the df subcommand which I fixed in a previous patch. The
fact that nobody had reported that in the time it existed
(probably several years) means that I'm rather doubtful that
anybody is using this tool anyway.

So the easiest solution seems to be removing the extra lines
of output.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>