fsck.gfs2: give comfort when processing lots of data blocks
When fsck.gfs2 is running, it can seem to silently hang when it is
processing a large number of data blocks.  This is noticeable when
files get over 20GB in size.  This patch periodically prints
status messages when big files are being checked so customers
know it's still doing something valid.

rhbz#455300