gfs2_convert: Fix conversion of inodes with different heights on gfs1 and gfs2
The first hunk of the patch fixes a possible typo where the gfs1 height was
being used instead of gfs2 height to compute the fan-out factors. When
both heights at gfs1 and gfs2 are the same, this really doesn't matter,
but at different heights, I get a floating point exception and gfs2_convert
crashes.

The second part of the fix has to do with updating the height information
of a given block after it's revised metapath is computed according to gfs2
parameters. Without this line, gfs2_convert was using the same height as that
of gfs1 for the converted file and weird blocks were included in the file's
data and metadata, thereby corrupting it.

Signed-off-by: Abhi Das <adas@redhat.com>