gfs2_quota: Bug 536902 - quota file size not a multiple of struct gfs2_quota
This is the userland portion of a patch-set that rips out the
linked list implementation in the quota file. It makes all
the quota linked list related warnings go away.
An accompanying kernel patch (tracked in bz 546455) is the
other half of this patch-set. This userland patch should
work harmoniously with an unpatched kernel, but it is
recommended that the kernel patch be applied as well.
Patched kernel and unpatched userland gfs2_quota is not
tested and is likely to crash.

It is also recommended that the quotas be reset on the
filesystem using "gfs2_quota reset -f /mnt/gfs2". This is to
avoid a possible corruption introduced by the linked-list
code.