Move duplicate code from libgfs2 to fsck.gfs2
This patch removes the duplicate block management from
libgfs2 and puts it into the only code that used it: fsck.gfs2.
This has several advantages for performance and readability.
Since it no longer has to worry about duplicates, the blockmap
code in libgfs2 can do less work in places where fsck.gfs2
doesn't care one way or another if the block is a duplicate.
Conversely, the places in fsck.gfs2 that only care if the
block is a duplicate are also simplified because they don't
have to do the additional blockmap code that went along.
This also paves the way for future code simplification.

rhbz#455300