gfs2: libgfs2 and fsck.gfs2 cleanups
This patch is big, but mostly cosmetic.  There are a few places
where I found and fixed a bug with the previous patches.  However,
by and large, this patch is comprised of (1) comment changes,
(2) code reformatting for readability, (3) variable renames to
aid in readability, (4) code restructuring for efficiency, as in
the case where I got rid of useless function scan_meta and just
inlined it, (5) improved error messages given to aid in debugging.

rhbz#455300