libgfs2: Set block range based on rgrps, not device size
The functions that read in the rindex need to set the block
range properly, rather than basing it on the device size.
If not, bad block pointers might reference blocks beyond the
end of the file system (as can happen if a device is extended
before gfs2_grow is run), or if a different metadata set is
restored with gfs2_edit restoremeta.

rhbz#455300