fsck.gfs2: cleanup: refactor pass3
This patch just cleans up pass3 a bit.  It is mostly just rearranging
the code to make it smaller and more readable.  A check that was at
the bottom of the pass3 function was moved to the top of the function
and a continue inserted, which made it possible to change the indentation
to be more readable.  I also removed a place where the function was
marking the dinode as modified.  The code should only mark it modified
in places where it actually IS modified, so this was removed.

rhbz#455300