fsck.gfs2: invalidate invalid mode inodes
When fsck.gfs2 encountered a dinode that had an invalid type
(file, directory, socket, etc), it marked that block in the blockmap
as an invalid block.  That caused pass2 to delete it AND its metadata.
Since the metadata had not been previously processed, it would be
deleted wholesale.  That's okay for most cases, but if there was
a duplicate reference to any of those metadata blocks, pass2 would
not realize it and free the block anyway, out from under the valid
reference.  In other words, the valid reference to the block would
remain intact but the block was freed in the bitmap because it was
associated with the dinode that had an unknown type.  That means big
time corruption because future block allocations could then reuse the
freed metadata, once again causing another duplicate block reference.

This patch invalidates the metadata associated with dinodes that
do not have a valid inode type.  That causes all of its metadata to
be processed and duplicate references are put on the new invalid
references list.  Later, in pass1b, the duplicate block references
are resolved and the last remaining reference is left in a valid
state.  And pass2 had no more cause to free the block.

rhbz#455300