fsck.gfs2: Don't add extended attrib blocks to list twice
This patch adds a simple check to pass1 where it is adding a
block to the list of blocks that contain an extended attribute.
If the block is already on the list, no need to add it again.
This can sometimes happen in duplicate reference situations and
other obscure code paths.

Since pass1 processes the file system sequentially by block number
it's enough to check against the block most recently added to the
list.  This is a performance increase; early versions of this
code ran the entire linked list of special blocks.  If the file
system was labelled for selinux, it could have millions of entries
on the linked list, causing an enormous slowdown.

rhbz#455300