fsck.gfs2: small parameter passing optimization
This patch is a small optimization in a very often-used function,
done for performance purposes.  Function handle_di in pass1 derives
the block number from the buffer already passed in rather than
having the block number passed in as well, and wasting time
pushing it to the stack and off the stack.

rhbz#455300