gfs2_edit savemeta: Don't release indirect buffers too soon
This patch fixes a regression in the "gfs2_edit savemeta" command
whereby system files such as journals were not being saved.
The function that processes indirect block lists was releasing
buffers in a fashion that works okay with the old buf.c
assumptions but too soon for the new buf.c assumptions.

rhbz#455300