cman: gfs_controld dm suspend hangs withdrawn GFS file system
This patch changes gfs_controld such that if a device mapper (dm)
suspend needs to be done, it passes the --nolockfs and --noflush
parameters.  Now that gfs and gfs2 are using the standard means
of calling freeze and thaw, a gfs withdraw will cause the
suspend uevent.  Without those parameters, the file system is
permanently frozen with no way to unfreeze.  (You can't unfreeze
it since gfs has withdrawn).  In other words, a reboot of the
entire cluster it required to recover.

rhbz#571806