gfs2_convert: gfs2_convert doesn't convert quota files
With this patch, gfs2_convert now assigns the metadata
and data blocks of the old quota file to the newly
created quota inode in the master directory.
Instead of having a initialized quota file, we now have
a quota file that preserves all the quota information
from the gfs1 filesystem.

Resolves: rhbz#589823
Signed-off-by: Abhi Das <adas@redhat.com>