Make gfs2_edit show bit-to-block translation when viewing bitmaps
This patch makes the gfs2_edit utility show the blocks and block
allocation values associated with the highlighted byte when walking
through the rgrp bitmaps in hex mode.  This means you can simply
walk your cursor to any given byte of any given bitmap and see
which four blocks are represented by that byte in the bitmap, and
whether each of them is free, data, metadata, or unlinked.