GFS2: Clean up initscript
The initscript for GFS2 appeared to have been set up with macros
to be filled in by the build system. The build system doesn't
actually do that, though.

This renames the initscript back to its original name and removes
the two macros which were included. Since it is by definition a
script, we can just add a line or two to check for files being
in different places if required. There is no need to do that via
the build system.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>