gfs2_edit: tiny (stuffed) files had user data saved with savemeta
The gfs2_edit savemeta option was saving the entire block for dinodes.
That's okay for average files where you need to save indirect block
pointers.  It's okay for directories where you need to save the
directory block pointers.  It's okay for system files where you want
to save statfs data or inum data and such.  But it's not okay with
stuffed user files because they contain user data.  This patch checks
for this special dinode condition and saves only the dinode
structure.

rhbz#634623