gfs_controld: Allow mounts entirely via sysfs/uevent interface
This patch adds the ability for gfs_controld to optionally allow
mounts via sysfs/uevents rather than requiring the mount.gfs2
helper. Some kernel support is required to do that, and the support
is in recent kernels (2.6.36 and up) so this has to be backwards
compatible.

The other change is that when a withdraw occurs, the device is now
specified by device number. This number is obtained from sysfs
and avoids the issue which would otherwise occur with the new
mount sequence for which the device path is not available.

There is a flag added to the mount group which determines whether
the mount helper exists or not. The flag is set depending on
whether the first notification of the filesystem is from the
ADD uevent or from the mount helper.

The plan is that the mount helper will be gradually phased out
over time.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>