gfs_controld: Allow paths names to be changed, but keep defaults
This should solve the problem which Fabio pointed out with making
the path names constant for the time being. The locations are
still changable, but there are also defaults in the code as well.

Also fixed a couple of warnings in the code which I spotted at the
same time.

Longer term we need to look carefully into whether we need to retain
either of the files in question, anyway. There are other ways of
ensuring that only one copy of a program runs at once - one way is
to open a unix socket of the type which gfs_controld already opens.

Also, we currently have three methods of logging (a) via the gfs_control
program and an internal buffer (b) to a log file and (c) to syslog.

Bearing in mind that (a) and (c) seem most useful, getting rid of (b)
seems like a possibility for the future.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>