build: New build system
This is carefully designed so that it can run alongside the existing
build system. It is experimental at the moment, so don't use it for
building packages just yet. I'm checking it in so that its easier to
work on it and to give a heads up on what I'm planning for the future.

Everything is run from the top level Makefile.new. When it is ready
to be used for real, I'll rename it to "Makefile" and we can then remove
the old build system.

Also added in this commit is a .pot file. Again please don't use it
(i.e. don't start adding translations) just yet. I've added it so we
can see what the strings file will look like, and thus we should be
able to tidy up the strings so that they are ready for translation
at a later date. The file can be regenerated with:

make -f Makefile.new strings

when required. When it is generated we get a few warnings regarding
the use of '\r' in strings, so we need to sort that out at least
before we can ask people to start translation work.

Not all the programs contribute to the .pot file. That is deliberate
since we don't want to include utilities which are obsolete and likely
to be dropped in future.

The new README.build file contains instructions on how to build the
source as well as a list of libraries required.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>