tune: Clean up and make closer to tune2fs
The -l messages have been reordered so that they now print
in the same order as tune2fs where the messages are common
and the messages have also been reformatted so that the
printed text is the same in the common cases.

The test for gfs2 has been corrected - the magic number is the
same for gfs and gfs2, we have to look at the format number to
tell the two versions apart.

The tool now prints the root and master inode numbers (for gfs2).

The strings have also been updated.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>