GFS2: mkfs.gfs2 segfaults with 18.55TB and -b512
The problem was that mkfs.gfs2 attempted to keep the number of resource
groups as low as possible, but in so doing, it chose a rgrp size that
required an enormous number of bitmap blocks per rgrp.  In fact, it
tried to use more than it could possibly address, given the small
block size.  Therefore, an error was flagged and mkfs.gfs2 aborted.
This patch ensures that mkfs.gfs2 chooses a rgrp size that takes into
account the maximum number of bitmap blocks.  If it hits the max
number of bitmap blocks, it backs off on its rgrp size and uses a
smaller size.  The smaller size rgrp ensures that a valid number of
bitmap blocks will be needed to represent all the blocks in each rgrp.

rhbz#624535