gfs2_convert: exits with success without doing anything
A recent gfs2-utils patch  added a sanity check (gfs2_check_range())
to the codepath of gfs1_ri_update() that's used by gfs2_convert.
This new check requires the value of sbp->fssize to be initialized
to a valid non-zero size, which gfs2_convert does not do. The check
fails and gfs2_convert bails out. This patch initializes sbp->fssize
to the device size and the check is successful and gfs2_convert runs
to completion as expected.

Resolves: rhbz#688734
Signed-off-by: Abhi Das <adas@redhat.com>