mount.gfs2 - gfs2 mounts doubled up in mtab
When the -o remount option is used with mount.gfs2 it fails to
remove the original mtab entry because it can't find it. Instead
it simply adds the mount entry corresponding to the remount,
thereby doubling the entries in mtab. This patch corrects the
logic to find the mtab entry so mount.gfs2 finds and removes
the original entry correctly before adding the remount entry.

Resolves: rhbz#706141
Signed-off-by: Abhi Das <adas@redhat.com>