fsck.gfs2: don't free prev rgrp list repairing rgrps
In cases where fsck.gfs2 is trying to repair damaged resource groups or
rindex, it tries several levels of repair.  Some of those levels build a
list of expected resource groups and check them against the actual ones,
and the previous list is freed before the next level is attempted.
However, in the case of minor damage, such as one bitmap block that was
overwritten, we can often repair the damage by reading in the resource
groups at the first level, then at the second level of repair (called
"ye_of_little_faith") use those values to compare against.  This patch
allows the second level of repair to use the first level's list.

rhbz#675723