fsck.gfs2: Rearrange block types to group all inode types
This patch rearranges the declares in util.h such that all the inode
types are grouped together.  There's a place in fsck.gfs2 that checks
whether a block is an inode by checking the range of block types.
But that's only valid if they are all grouped together.  Since a
previous patch put GFS journaled data blocks in the range, they could
mistakenly pass that test and be treated as an inode for that test.
I never saw an error resulting from this mistake, but I'm fixing it
anyway.

rhbz#675723