gfs2_edit: Fix segfault jumping within rindex
This patch fixes a segfault that occurs when you jump to an rindex
entry from the extended mode.  The problem was that the jump code
was assuming jumps occurred only from structures that had an indirect
data structure.  The rindex doesn't, so it tried to reference an
invalid entry in the indirect structure.