libgfs2: Fix off-by-one error in rgrp searching
The patch fixes an off-by-one error that results in landing on the
wrong resource group when searching for an address exactly equal to
an rgrp's ri_addr.

This bug is mostly seen if you're walking the rgrps from gfs2_edit
and walk the bitmaps: The bitmap's block translation will be
calculated for the wrong rgrp.  The code was checking to see if the
block is greater than ri->ri_data0 + ri->ri_data, but technically
ri->ri_data0 + ri->ri_data is the ri_addr of the next rgrp.
So the correct calculation should be block > ri_data0 + ri_data - 1
or for simplicity of coding, block >= ri_data0 + ri_data