libgfs2: Improve rgblocks2bitblocks()
This patch reworks the rgblocks2bitblocks function which was
inefficient, difficult to read and generally unwieldy.

As this is core code from the days of yore and fsck.gfs2 depends on it,
I made sure to test the new function extensively, comparing its outputs
with the original function over a large range of values for rgblocks (up
to 195312500) and valid block sizes between 512 and 4096.

All call points have been updated and, as a nice side effect, the run
time of the function is greatly reduced.

Signed-off-by: Andrew Price <anprice@redhat.com>