libgfs2: Prepare to remove log_xxx() macros from library
This will need to be done in stages. The first stage is to
clean up what is there. There are calls to the log_ functions
in the library itself, and those will need to be removed
in later patches.

This patch moves the gfs2_options structure into the two
utils which actually require it, removes the print_fsck_log
function, replacing the 3 calls to it from fsck with
the equivalent printf functions - it looks like those could
be merged in future as they all look very similar.

The log_ macros are also updated, so that they will do excatly
the same as before, but now don't depend on any code in libgfs2
itself.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>