libgfs2: Clean up sb read/check functions
There is no need to specify in advance whether we want to accept
gfs1 format super blocks or not. We can simply return that info
and let the caller decide if gfs1 format is acceptable or not.

We can also remove the message regarding the magic number of the
superblock, since the caller prints out a message anyway if
the check fails.

That removes some more of the log_ messages from libgfs2 as a
result.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>