libgfs2: Clean up device geometry code
The main part of this patch is the creation of a new structure
which will contain all the information that we can discover
about a device. Currently this means a struct stat plus the
results of various ioctls, and the size. A device in this case
might also be a regular file, of course.

Since all the code does is gather information, we can use the
same code in all the utilities.

The existing struct device had two fields which were unused
or almost entirely unused. Those have been removed, greatly
simplifying the code in that area.

At the same time this code also fixes a bug in the device
topology ioctls (the alignment offset and optimal i/o size
were switched in the original code).

I've done a few quick tests and all appears to be working
correctly. There should be no noticeable changes in
behaviour to the user. All the calculations should give
the same results, both before and after (aside from the ones
affected by the topology bug).

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>