libgfs2: Remove more unused macros/functions
Two macros in lingfs2.h, bsync and bcommit referred to functions
which no longer exist, so this removes them.

The __bget_generic function was exported, but only ever used
internally to libgfs2, so that is unexported and split
between the two functions which called it, since both had
differing requirements, anyway. This also means that we
can remove the bget wrapper for __bget since this part made
no use of the passed in function name and line number
anyway.

There should be no resulting change in function after this
patch, but another 20 lines of unused code has been removed.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>