gfs2_convert: clean up question asking code
The convert utility doesn't need lots of fancy processing when
it asks the question of whether to proceed or not. As a result
most of the code thats there can be removed.

One thought that occurred to me was whether we should block
SIGINT like we do with fsck, during operation. That can be
a separate patch though, and it would be much simpler since there
is no reason to prompt the user as to whether to continue or not
in the conversion case, since a partial conversion would be useless.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>