libgfs2: Add beginnings of a metadata description
The intent here is to start a framework which will describe the
GFS/GFS2 metadata in enough detail to be used by the utilities.
Initially there is little more than names, sizes and offsets of
the various fields, but I hope to expand on this a bit in future.

The other thing that I've added, is a simple self-check function.
Again, I hope that we can build on this base and expand the tests
as time goes on.

Currently nothing uses this new infrastructure, but again, in time
it should be possible to integrate it into the utilities. In the
mean time I want to make sure it is as complete and correct as
possible.

A side effect of this is that every bit of metadata has been
assigned a unique id. That includes those items which do not
have the gfs2_meta_header at the start of them, and so don't
have a type or format number.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>