libgfs2: Remove struct gfs_rindex from header, etc
Since struct gfs_rindex is identical to gfs2_rindex, we don't
really need two structures. It was used in two places. Once
to define an unused member of a union with a gfs2_rindex and
once to define a buffer size (which I think was a mistake anyway
since the buffer contains a struct gfs2_rindex anyway.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>