libgfs2: Use endian defined types for GFS1 on disk structures
Since we are using these types anyway, for the GFS2 on disk
structures, and we are already using some of those as parts
of the GFS on disk structures (e.g. the metadata header) then
lets be consistent in the types we are using.

Also, this has the advantage that sparse will underastand them too.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>