libgfs2: Set errno in dirent_alloc and use dir_add consistently
Initially spotted by coverity: "error" is passed to a parameter that
cannot be negative.

It turned out that dirent_alloc was returning -ENOSPC and some calling
code (via dir_add) was assuming that errno was set. Other code was
assuming that dir_add returned -errno. This patch changes dirent_alloc
to set errno and return -1 and also updates the code which calls dir_add
to use errno.

Signed-off-by: Andrew Price <anprice@redhat.com>