mkfs.gfs2: Improve error messages
Currently mkfs.gfs2 does not warn that the lock table has been omitted,
it just complains about a missing colon in the lock table. It also
prefixes error messages with the absolute path of the source file. This
patch removes the path from error messages (in mkfs and edit) and adds a
check for an empty lock table.

Signed-off-by: Andrew Price <anprice@redhat.com>