libgfs2: Add flags to metadata description
This patch adds a flags field to the metadata description so that we can
mark up the properties of certain fields. This can also be expanded in
due course to accommodate further information if required.

Also, this adds the missing jdata structure into the metadata
description. I think every structure should be there now.

Structures which consisted only of a single metaheader have been given
their own structures so we can use the MH macro and keep the naming
consistent across all the structures.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>