libgfs2: Add pointer restriction flags
This patch adds a couple of new metadata block types to represent free
blocks and data blocks. This then allows these types to be used to
specify all the valid block target types which pointers might point to.
We can then add a new field to store this restriction information.

Also this adds a new test to check that the restriction information is
supplied for all pointers.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>