gfs2_edit restoremeta: sync changes on a regular basis RHEL6
This patch calls fsync on a regular basis to commit restored
GFS2 metadata to media. This is primarily to make it more user-
friendly. Without it, large metadata sets appear to restore very
quickly, then hang for a very long time while the buffers are all
being synced to disk. This prevents it from getting too far ahead.
So it will seem to restore more slowly, but not hang at the end.