gfs2_utils: Improve error messages
Currently gfs2_edit complains that it doesn't know what 'master' means
when it can't find the master directory. This patch catches the case
where the master directory hasn't been found and reports the error more
clearly to the user. It also updates a terse 'out of space' error
message in blk_alloc_i to be more clear.

rhbz#742293

Signed-off-by: Andrew Price <anprice@redhat.com>