libgfs2: Add a gfs2 block query language
This patch adds a small language which we can use to write libgfs2 and
gfs2-utils tests (among other things). It provides 'get' and 'set'
statements which look up and modify gfs2 blocks. The language API is
defined as:

   struct lgfs2_lang_state;

   struct lgfs2_lang_result {
           uint64_t lr_blocknr;
           struct gfs2_buffer_head *lr_bh;
           const struct lgfs2_metadata *lr_mtype;
           int lr_state; // GFS2_BLKST_*
   };

   struct lgfs2_lang_state *lgfs2_lang_init(void);
   void lgfs2_lang_free(struct lgfs2_lang_state **state);

   int lgfs2_lang_parsef(struct lgfs2_lang_state *state, FILE *script);
   int lgfs2_lang_parses(struct lgfs2_lang_state *state, const char *script);

   struct lgfs2_lang_result *lgfs2_lang_result_next(struct lgfs2_lang_state *state, struct gfs2_sbd *sbd);
   int lgfs2_lang_result_print(struct lgfs2_lang_result *result);
   void lgfs2_lang_result_free(struct lgfs2_lang_result **result);

The lgfs2_lang_parse{s,f} functions allow you to parse a string or a
file respectively. Using the same state object you can then run the same
script multiple times, without parsing it again. The intended usage of
these functions can be shown with a simple example (error checking
omitted):

   struct lgfs2_lang_state *state = lgfs2_lang_init();
   lgfs2_lang_parses(state, "get sb; get 1234 state; set '/foo/bar' {di_entries: 3}");
   for (result = lgfs2_lang_result_next(state, &sbd);
        result != NULL;
        result = lgfs2_lang_result_next(state, &sbd)) {
           lgfs2_lang_result_print(result);
           lgfs2_lang_result_free(&result);
   }
   lgfs2_lang_free(&state);

The language has a simple syntax:

   get <block_lookup> [state]
   set <block_lookup> {<field0>:<value0>, <field1>:<value1>, ... }

A block lookup can be a file system block address (1234 or 0x1234), a
resource group subscript (rgrp[0], rgrp[0x5]), a keyword (sb, master,
root, rindex) or an offset from any of the above (1234+5, rgrp[1]+23,
rindex+68, ...).

After the block lookup, the 'get' statement takes an optional keyword
'state' which allows you to query the bitmap state of the block.

Examples of 'get':

   get rgrp[1]+23 state
   --> result.lr_state == GFS2_BLKST_FREE

   get rindex
   --> result.lr_bh is a buffer containing the rindex dinode.
   --> result.lr_mtype is the block's metatype as defined in meta.c.

The 'set' statement requires a list of field-value pairs which are to be
modified in the block. The field names must match the names shown when
running the 'get' statement with the same block lookup.

Examples of 'set':

   set '/foo/bar/baz' { di_entries: 3 }
   --> result.lr_bh contains the dinode block at path /foo/bar/baz, with
       the di_entries field modified.
   --> result.lr_mtype is the block's metatype as defined in meta.c.

   set sb { sb_lockproto: 'lock_dlm', sb_bsize: 0x1000 }
   --> result.lr_bh contains the superblock with the lockproto changed to
       lock_dlm and the block size changed to 4096.
   --> result.lr_mtype is the block's metatype as defined in meta.c.

Whitespace is insignificant in the language and semicolons can be used
to separate multiple statements. When writing longer scripts spanning
multiple lines, C-like // syntax can be used to insert comments.

Signed-off-by: Andrew Price <anprice@redhat.com>