libgfs2: Move valid_block into fsck.gfs2
valid_block() is only used by fsck.gfs2 as none of the other utils
require the same level of paranoia so we can move it into fsck.h and
make it static inline.

In my (small) tests the speedup effect is barely noticeable but it
should make a difference for larger file systems, particularly those
with very large directories.

Also fsck/fs_bits.h is no longer used so this patch removes it.

Signed-off-by: Andrew Price <anprice@redhat.com>